Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -2,7 +2,7 @@
 
 
 int main() {
-    const pie=3.14;
+    const (float) pie=3.14;
     float a;
     float num;
     num=a*a;
